sc - refactor id resolution into a util file
